+++ src/main/java/coffeeshop/Models/DAO.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 public interface DAO<Type> {
-    Optional<Type> get(long id);
+    Optional<Type> get(Integer id);
     List<Type> getAll();
     void save(Type t);
     void update(Type t);
